finish jiff deprecations
