+++ python/ry/ryo3/_jiff.pyi
@@ -308,10 +308,6 @@ class Date(
     def iso_week_date(self) -> ISOWeekDate: ...
     def in_leap_year(self) -> bool: ...
     def in_tz(self, tz: TimezoneName) -> ZonedDateTime: ...
-    @deprecated(
-        "`Date.intz` is deprecated; use `Date.in_tz` instead [removal: v0.0.93]"
-    )
-    def intz(self, tz: TimezoneName) -> ZonedDateTime: ...
     def last_of_month(self) -> t.Self: ...
     def last_of_year(self) -> t.Self: ...
     def nth_weekday(self, nth: int, weekday: Weekday) -> t.Self: ...
@@ -715,10 +711,6 @@ class DateTime(
     def first_of_year(self) -> t.Self: ...
     def in_leap_year(self) -> bool: ...
     def in_tz(self, tz: TimezoneName) -> ZonedDateTime: ...
-    @deprecated(
-        "`DateTime.intz` is deprecated; use `DateTime.in_tz` instead [removal: v0.0.93]"
-    )
-    def intz(self, tz: TimezoneName) -> ZonedDateTime: ...
     def iso_week_date(self) -> ISOWeekDate: ...
     def last_of_month(self) -> t.Self: ...
     def last_of_year(self) -> t.Self: ...
@@ -1523,11 +1515,6 @@ class Timestamp(
     def as_second(self) -> int: ...
     def display_with_offset(self, offset: Offset) -> str: ...
     def in_tz(self, tz: TimezoneName) -> ZonedDateTime: ...
-    @deprecated(
-        "`Timestamp.intz` is deprecated; use `Timestamp.in_tz` instead [removal: v0.0.93]"
-    )
-    def intz(self, tz: TimezoneName) -> ZonedDateTime:
-        """Deprecated ~ use `in_tz`"""
     @property
     def is_zero(self) -> bool: ...
     def series(self, period: TimeSpan) -> JiffSeries[t.Self]: ...
@@ -1786,10 +1773,6 @@ class ZonedDateTime(
     def first_of_year(self) -> t.Self: ...
     def in_leap_year(self) -> bool: ...
     def in_tz(self, tz: TimezoneName) -> t.Self: ...
-    @deprecated(
-        "`ZonedDateTime.intz` is deprecated; use `ZonedDateTime.in_tz` instead [removal: v0.0.93]"
-    )
-    def intz(self, tz: TimezoneName) -> t.Self: ...
     def inutc(self) -> t.Self: ...
     def last_of_month(self) -> t.Self: ...
     def last_of_year(self) -> t.Self: ...
